Enforce that embeddings are normalized to 1 and use float32 precision for embeddings
With the OpenAI API, embeddings are already normalized to 1, but now the embeddings are normalized to 1 regardless, to generalize to cases in which other embeddings are used. It was observed that changing the embedding precision to float32 gives the same results as float64, so float32 is used.

+++ src/embeddings.py
@@ -25,7 +25,7 @@ def create_embeddings(texts):
             input=texts
         )
         embeddings = [e.embedding for e in response.data]
-        return np.array(embeddings)
+        return np.array(embeddings, dtype=np.float32)
     except Exception as e:
         raise RuntimeError(f"Error while generating embeddings: {e}")
